requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
requests
numpy
annoy
python-multipart
streamlit
openai
pydantic
tqdm
python-dotenv
chromadb
pdfminer.six
docx
python-docx
